moonbase_runtime
Type Alias ParachainStakingConfig

pub type ParachainStakingConfig = GenesisConfig<Runtime>;

Aliased Type

struct ParachainStakingConfig {
    pub candidates: Vec<(AccountId20, u128)>,
    pub delegations: Vec<(AccountId20, AccountId20, u128, Percent)>,
    pub inflation_config: InflationInfo<u128>,
    pub collator_commission: Perbill,
    pub parachain_bond_reserve_percent: Percent,
    pub blocks_per_round: u32,
    pub num_selected_candidates: u32,
}

Fields
candidates: Vec<(AccountId20, u128)>

Initialize balance and register all as collators: (collator AccountId, balance Amount)

delegations: Vec<(AccountId20, AccountId20, u128, Percent)>

Initialize balance and make delegations: (delegator AccountId, collator AccountId, delegation Amount, auto-compounding Percent)

inflation_config: InflationInfo<u128>

Inflation configuration

collator_commission: Perbill

Default fixed percent a collator takes off the top of due rewards

parachain_bond_reserve_percent: Percent

Default percent of inflation set aside for parachain bond every round

blocks_per_round: u32

Default number of blocks in a round

num_selected_candidates: u32

Number of selected candidates every round. Cannot be lower than MinSelectedCandidates