Type Alias moonbase_runtime::xcm_config::XcmFeesToAccount

pub type XcmFeesToAccount = XcmFeesToAccount<Assets, (ConvertedConcreteId<AssetId, Balance, AsAssetType<AssetId, AssetType, AssetManager>, JustTry>,), AccountId, XcmFeesAccount>;

This is the struct that will handle the revenue from xcm fees We do not burn anything because we want to mimic exactly what the sovereign account has

Aliased Type

struct XcmFeesToAccount(/* private fields */);