Type Definition moonbase_runtime::xcm_config::XcmOriginToTransactDispatchOrigin

pub type XcmOriginToTransactDispatchOrigin = (SovereignSignedViaLocation<LocationToAccountId, Origin>, RelayChainAsNative<RelayChainOrigin, Origin>, SiblingParachainAsNative<Origin, Origin>, XcmPassthrough<Origin>, SignedAccountKey20AsNative<RelayNetwork, Origin>);

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.