Module moonbase_runtime::xcm_config

XCM configuration for Moonbase.

Structs

BaseXcmWeight

CurrencyIdToLocation

Erc20XcmBridgePalletLocation

Erc20XcmBridgeTransferGasLimit

EvmRunnerPrecompileOrEthXcm

LocationToH160

Wrapper type around LocationToAccountId to convert an AccountId to type H160.

MaxAssetsForTransfer

MaxAssetsIntoHolding

MaxHrmpRelayFee

MaxInstructions

Maximum number of instructions in a single XCM fragment. A sanity check against weight caculations getting too crazy.

MessageQueueHeapSize

The size of the page; this implies the maximum message size which can be sent.

MessageQueueMaxStale

The maximum number of stale pages (i.e. of overweight messages) allowed before culling can happen. Once there are more stale pages than this, then historical pages may be dropped, even if they contain unprocessed overweight messages.

MessageQueueServiceWeight

The amount of weight (if any) which should be provided to the message queue for servicing enqueued items.

MoonbeamCall

ParachainMinFee

RelayChainOrigin

RelayNetwork

RelayOrigin

SafeCallFilter

SelfLocation

SelfLocationAbsolute

SelfReserve

UniversalLocation

XcmExecutorConfig

XcmFeesAccount

Xcm fees will go to the treasury account

Enums

AssetType

CurrencyId

Transactors

Type Aliases

AssetTransactors

DerivativeAddressRegistrationOrigin

ForeignFungiblesTransactor

LocalAssetTransactor

The transactor for our own chain currency.

LocalOriginToLocation

LocationToAccountId

Type for specifying how a Location can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

XcmBarrier

XcmFeesToAccount

This is the struct that will handle the revenue from xcm fees We do not burn anything because we want to mimic exactly what the sovereign account has

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.

XcmWeigher

Xcm Weigher shared between multiple Xcm-related configs.